+++ src/jp/crestmuse/cmx/handlers/SCCHandler.java
@@ -5,7 +5,7 @@ import jp.crestmuse.cmx.filewrappers.*;
 public interface SCCHandler {
   public void beginHeader(SCCXMLWrapper w);
   public void endHeader(SCCXMLWrapper w);
-  public void processHeaderElement(int timestamp, String name,
+  public void processHeaderElement(long timestamp, String name,
                                    String content, SCCXMLWrapper w);
   public void beginPart(SCCXMLWrapper.Part part, SCCXMLWrapper w);
   public void endPart(SCCXMLWrapper.Part part, SCCXMLWrapper w);
